Edited delete to account for guild vs global

+++ commands/delete.js
@@ -5,18 +5,19 @@ const colours = require("../colours.json");
 module.exports = {
     name: "delete",
     description: "Moderator tool to delete slash commands",
-    usage: "?delete [slash command id]",
+    usage: "?delete [slash command id] [optional: delete guild command? Leave blank if false]",
     async execute(bot, message, args) {
 
         if (!message.member.hasPermission(["ADMINISTRATOR"])) return message.reply("You don't have that permission");
         if (!args[0]) return message.reply("No id given for slash command.")
         let id = args[0]
         try {
-            if (botconfig.GUILD_ID) {
+            if (botconfig.GUILD_ID && args[1] == 'true') { //args[1] true if Guild Testing Command deletion
                 console.log("In guild server")
                 await bot.api.applications(bot.user.id).guilds(botconfig.GUILD_ID).commands(id).delete()
             } else {
-                await bot.api.applications(bot.user.id).commands(id).delete()
+                console.log("Not deleting guild command")
+                //await bot.api.applications(bot.user.id).commands(id).delete()
             }
 
             return message.reply('Slash command has been deleted')
